Some compilation fixes (uninitialized but used variables, unused variables, etc.) discovered when compiling with clang.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a code cleanup that fixes compilation issues by initializing variables, adjusting data types, and removing unused variables to resolve ASAN and UBSAN warnings.

#### PR Summary
The changes improve code reliability by initializing previously uninitialized variables, correcting type mismatches, and eliminating unused variables.
- `acvp_testing.h`: Updated test data length declarations from `uint32_t` to `size_t`.
- `flash_master_mock.c`: Replaced a for loop with a while loop to remove an unused loop index.
- `key_cache_flash.c`: Added an explicit cast for the key status to match the expected enum type during cache initialization.
- `spdm_secure_session_manager.c`: Initialized `local_pub_key_der_len` to 0 to prevent uninitialized variable warnings.
- Test files (`logging_flash_test.c` and `ephemeral_key_manager_test.c`): Initialized variables and added assertions to validate status, ensuring test robustness.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #5504

+++ core/testing/acvp/acvp_testing.h
@@ -7,22 +7,22 @@
 
 /* Sample ACVP test input and output data. */
 extern const uint8_t ACVP_PROTO_TEST_DATA_SHA256[];
-extern const uint32_t ACVP_PROTO_TEST_DATA_SHA256_LEN;
+extern const size_t ACVP_PROTO_TEST_DATA_SHA256_LEN;
 
 extern const uint8_t ACVP_PROTO_TEST_DATA_SHA384[];
-extern const uint32_t ACVP_PROTO_TEST_DATA_SHA384_LEN;
+extern const size_t ACVP_PROTO_TEST_DATA_SHA384_LEN;
 
 extern const uint8_t ACVP_PROTO_TEST_DATA_SHA512[];
-extern const uint32_t ACVP_PROTO_TEST_DATA_SHA512_LEN;
+extern const size_t ACVP_PROTO_TEST_DATA_SHA512_LEN;
 
 extern const uint8_t ACVP_PROTO_TEST_RESULTS_SHA256[];
-extern const uint32_t ACVP_PROTO_TEST_RESULTS_SHA256_LEN;
+extern const size_t ACVP_PROTO_TEST_RESULTS_SHA256_LEN;
 
 extern const uint8_t ACVP_PROTO_TEST_RESULTS_SHA384[];
-extern const uint32_t ACVP_PROTO_TEST_RESULTS_SHA384_LEN;
+extern const size_t ACVP_PROTO_TEST_RESULTS_SHA384_LEN;
 
 extern const uint8_t ACVP_PROTO_TEST_RESULTS_SHA512[];
-extern const uint32_t ACVP_PROTO_TEST_RESULTS_SHA512_LEN;
+extern const size_t ACVP_PROTO_TEST_RESULTS_SHA512_LEN;
 
 
 #endif	/* ACVP_TESTING_H_ */
